requirements.txt
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
openai==1.51.0
markdownify==0.13.1
